Add classifyElevationOnGrid to DEM and switch watershed elevation statistics to double precision
Introduce ReosDigitalElevationModel::classifyElevationOnGrid and its implementation for raster DEMs to group DEM values by byte-class grid cells (returns a QVector of per-class float lists). Update API headers accordingly.

Use QVector> for the classified elevation container in watershed delineation and switch stored/calculated per-class mean and standard deviation from float to double (accumulation and accessors updated). Minor cleanup: remove an extraneous semicolon in ReosHydraulicNetwork header.

These changes allow classifying elevations by distance classes and compute more numerically robust statistics for watershed delineation.

+++ src/core/watershed/reoswatersheddelineating.h
@@ -67,8 +67,8 @@ class ReosWatershedDelineatingProcess : public ReosProcess
     double averageElevation() const;
     QVector<int> distanceArea() const;
     QList<int> areaToElevationCount() const;
-    QList<float> areaToElevationMean() const;
-    QList<float> areaToElevationStd() const;
+    QList<double> areaToElevationMean() const;
+    QList<double> areaToElevationStd() const;
 
     bool calculateAverageElevation() const;
 
@@ -84,8 +84,8 @@ class ReosWatershedDelineatingProcess : public ReosProcess
     ReosRasterWatershed::DistanceClasses mDistanceClasses;
     QVector<int> mDistanceToArea;
     QList<int> mAreaToElevationCount;
-    QList<float> mAreaToElevationMean;
-    QList<float> mAreaToElevationStd;
+    QList<double> mAreaToElevationMean;
+    QList<double> mAreaToElevationStd;
     QPolygonF mOutputWatershed;
     QPolygonF mOutputStreamline;
     ReosRasterExtent mPredefinedRasterExtent;
@@ -194,8 +194,8 @@ class REOSCORE_EXPORT ReosWatershedDelineating : public ReosModule
         double averageElevation;
         QVector<int> distanceArea;
         QList<int> areaToElevationCount;
-        QList<float> areaToElevationMean;
-        QList<float> areaToElevationStd;
+        QList<double> areaToElevationMean;
+        QList<double> areaToElevationStd;
     };
 
     static DelineateResult delineateWatershed(
